requirements_core.txt
Django==1.6.1
djangorestframework==2.3.12
South==0.8.4
django-compressor==1.3
django-filter==0.6
Pillow==2.3.0
dj-database-url==0.2.1
django-storages==1.1.8
boto==2.9.6
django-import-export=0.2.1